+++ packages/moonstone/CHANGELOG.md
@@ -7,6 +7,8 @@ The following is a curated list of changes in the Enact moonstone module, newest
 ### Deprecated
 
 ### Added
+- `moonstone/VideoPlayer` added ability to seeking when holding down the right and left keys. Can also adjust sensitivity using throttling options `jumpDelay` and `initialJumpDelay`.
+- `moonstone/VideoPlayer` added `no5wayJump` to disable jumping and holding done by 5-way
 
 - `moonstone/VideoPlayer` support for the "More" button to use tooltips
 - `moonstone/VideoPlayer` properties `moreButtonLabel` and `moreButtonCloseLabel` to allow customization of the "More" button's tooltip and Aria labels